requirements.txt
django-cors-headers==4.9.0
django-ninja-jwt==5.4.4
django-ninja==1.6.2
Django==6.0.3
mssql-django==1.7
reportlab==4.4.10
waitress==3.0.2
